parsing colour geslaagd

+++ srcs/libft/ft_free_array.c
@@ -6,13 +6,15 @@
 /*   By: nsterk <nsterk@student.codam.nl>             +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/03/22 15:44:37 by nsterk        #+#    #+#                 */
-/*   Updated: 2021/03/22 15:51:38 by nsterk        ########   odam.nl         */
+/*   Updated: 2021/03/22 16:16:55 by nsterk        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 int	ft_free_array(char **str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i])
